document_processor/requirements.txt
python-docx
PyPDF2
reportlab
pyaspeller
pyspellchecker
textacy>=0.11
spacy>=3.0
